requirements.txt
fastapi
uvicorn
python-dotenv
asyncpg
Pillow
